learn_opentelemetry/python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "opentelemetry-api>=1.37.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.37.0",
    "opentelemetry-sdk>=1.37.0",
    "opentelemetry-instrumentation-logging>=0.41b0"
]
